Answer load on the open sockets
I also spent way too much time dreaming that I was using structlog and thus not catching that my functions were all breaking because logging wasn't behaving as I expected.

There seems to be radio silence when we first try to load code, as the supervisor does not accept logs during this period.